5.0 - Added getUserTimeStampFileName function in filepath
Summary:
This is the same as revision D6562 in od trunk.
As the requirement of PDF3D plugin, we need automatically generate the pdf file name based on timestamp. The new function is added and also replaces
part of code in debug.cc.

Test Plan: Compiles fine. Tests fine in show log file.

Reviewers: bert

Reviewed By: bert

Differential Revision: https://phabricator.dgbes.com:8443/D6589

git-svn-id: https://svn.opendtect.org/od/branches/od5.0@37593 a985855d-d384-4b45-ae33-31a410998c3c