od: fix for extra rounding off of numbers in scientific notation
Summary: A number like 1.599917e-5 was converted to the string "2e-5" leading to major issues while storing certain data (e.g. Horizons) in ASCII. The reason is that the number # in the formatting string "%.#f" specifies the number of digits after the decimal (as expected) but in "%.#g" it is the total number of significant digits. So, a "%.1g" formatting of 1.599917e-5 results in 2e-5.

Test Plan:
1. Now the number 1.599917e-5 converts nicely to 1.6e-5.
2. Added two more cases to test_string and the test succeeds.

Reviewers: kristofer, bert

Reviewed By: bert

Subscribers: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D6622

git-svn-id: https://svn.opendtect.org/od/trunk@37637 a985855d-d384-4b45-ae33-31a410998c3c