od: Added auto-correction of time-depth model before showing its editor
Summary: depth to time conversion of the track now contains a silent auto-correction if necessary. Used that auto-correction method in the construction of the object manager, so the user can write a corrected model to disk as well.

Test Plan: tested in F3_Demo on correct and incorrect model: well F2-01 contained two pairs of points with identical travel times.

Reviewers: bert

Reviewed By: bert

Subscribers: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D6625

git-svn-id: https://svn.opendtect.org/od/trunk@37642 a985855d-d384-4b45-ae33-31a410998c3c