Bugfix: Volume box disappeared while dragging along a dimension with survey step not unity (Mantis 3692). (PORTING FROM od TO od5.0 PRESERVING API AND OLD NAMING)
Summary:
Volume box disappears while dragging at the side of the survey origin along a
dimension with survey step not unity. Snapping was formerly done afterwards
in manipMotionFinishCB(). Should now be done already during draggerMoveCB() to
offer users WYSIWYG when using the box dragger to scroll through a data cube,
instead of an unexpected jump back when releasing the mouse. Volume disappeared
because I did not expect TrcKeyZSampling::limitTo(.) to return "undefined"
if two samplings with the same non-unit step are overlapping but not aligning.
Can't say it's wrong, it's what one defines it to be.

Test Plan:
Tested with the debug survey supplied by Yuancheng, which had non-unit step
crl- and z-dimensions. Volume no longer disappears.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D6639

git-svn-id: https://svn.opendtect.org/od/branches/od5.0@37651 a985855d-d384-4b45-ae33-31a410998c3c