od: Implemented comparison of geometries
Summary: This is useful while importing Seismic data. One can determine if the geometry of the data being imported (2D line or a 3D lattice) already exists in the database.

Test Plan: Tested by comparing the geometry from a segydef file and the ones in GeometryManager. The findRelated function throws the right geometry.

Reviewers: ranojay, bert

Reviewed By: bert

Subscribers: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D6649

git-svn-id: https://svn.opendtect.org/od/trunk@37661 a985855d-d384-4b45-ae33-31a410998c3c