od5.0: Fix for wrong order in exported 2D segy
Summary: While exporting 2D Seismics as SEGY, the 'writer' SeisTrcTranslator had is_2d flag 'false' causing a wrong sorting of traces before writing. The reason was that the Conn in this case was not linked to a real IOObj. This change uses the Conn->linkedTo() object only when it is valid. Otherwise the preset value of is_2d flag prevails.

Test Plan: Export of 2D lines gives correct order of traces in the SEGY file. Also debugging shows that is_2d flag in SeisTrcTranslator is 'true' while reading/writing 2D seismics.

Reviewers: bert

Reviewed By: bert

Differential Revision: https://phabricator.dgbes.com:8443/D6681

git-svn-id: https://svn.opendtect.org/od/branches/od5.0@37713 a985855d-d384-4b45-ae33-31a410998c3c