od5.0: Fix for removing current survey when aborting new survey creation
Summary:
the survey information should not be read when the survey editor was closed for a new survey creation: The survey has already been deleted. Reading the survey info from the file sets the current survey to the previous survey, which gets deleted.
http://bugs.dgbes.com/view.php?id=3704

Test Plan: no more deletion of current survey

Reviewers: nageswara, raman

Reviewed By: raman

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D6694

git-svn-id: https://svn.opendtect.org/od/branches/od5.0@37735 a985855d-d384-4b45-ae33-31a410998c3c