5.0- Re-Track All Bug 3714 ( The button will be removed from tool bar )
Summary:
This bug is caused by run "Re-Track All" in multi-times. The first time is OK. But the second time it will caused a dead loop between 
EMObjectDisplay::emChangeCB and HorizonDisplay::emChangeCB, because in the event: EM::EMObjectCallbackData::PositionChange, 
the triggermovement is set as true in EMObjectDisplay::emChangeCB.

After check the history and discuss with Nanne and Arnaud this function should be removed from 5.0.
See 4.6 history D30743

Test Plan: Compiles fine. Tests: the button is not displayed any more.

Reviewers: arnaud

Reviewed By: arnaud

Subscribers: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D6739

git-svn-id: https://svn.opendtect.org/od/branches/od5.0@37770 a985855d-d384-4b45-ae33-31a410998c3c