od5.0: Fixed a seeds based geo-body creation bug
Summary:
Bug ref: 2868
When set an output value for a position, we should add it as seed when the input value fits the threshold.

Previous code returned without adding proper seed

Test Plan: Tested on F3 and SEGSalt projects, was able to create bodies from a single or multiple seeds.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D6783

git-svn-id: https://svn.opendtect.org/od/branches/od5.0@37845 a985855d-d384-4b45-ae33-31a410998c3c