od5.0: bug fixing about randomdisplay name in restore session
Summary:
Although randomdisplay is one case, I would like to solve the problem on top level to avoid possible other objects having the same problem.
That is 4.6 does. The consequence is that the 2d line session file has also the object name. but I don't think that is a problem.

Test Plan: Compiles and Tests fine

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D6825

git-svn-id: https://svn.opendtect.org/od/branches/od5.0@37877 a985855d-d384-4b45-ae33-31a410998c3c