od: bug fixing about randomdisplay name in restore session
Summary:
This revision is same as D6825 in od5.0
Although randomdisplay is one case, I would like to solve the problem on top level to avoid possible other objects having the same problem.
That is 4.6 does. The consequence is that the 2d line session file has also the object name. but I don't think that is a problem.

Test Plan: Compiles and Tests fine

Reviewers: raman

Reviewed By: raman

Subscribers: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D6834

git-svn-id: https://svn.opendtect.org/od/trunk@37888 a985855d-d384-4b45-ae33-31a410998c3c