od:Change well text max setting as 500
Summary:
This is same as revision D6838 in od5.0
This requested from Farrukh who works on the project with a big bounding box. the text is displayed with the size related to bounding box size in osg.
It doesn't have influence on previous settings and projects.

Test Plan: Compiles and Tests fine

Reviewers: arnaud

Reviewed By: arnaud

Subscribers: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D6839

git-svn-id: https://svn.opendtect.org/od/trunk@37892 a985855d-d384-4b45-ae33-31a410998c3c