od: Wrong units in well tie window( Mantis 3752 )
Summary:
http://bugs.dgbes.com/view.php?id=3752
The problem here was that prior to V4.4 the key to store density and velocity 
log name was in reverse order. The key for velocity was used for density and vice versa.
This was fixed later on but still there are par files written in older versions.
The change is intended to support those older files. An alternative to this 
would be to ask the user to overwrite the existing WellTie Setup file with newer version.

Test Plan: Units are correvtly displayed in Well Tie Window.

Reviewers: raman, arnaud

Reviewed By: arnaud

Subscribers: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D6850

git-svn-id: https://svn.opendtect.org/od/trunk@37908 a985855d-d384-4b45-ae33-31a410998c3c