od: Fix for crash while writing Seismic Data when Z Range subselection has been done in Volume Builder
Summary:
Fix for Mantis Issue 3767, 3772.
In DataCubesWriter::nextStep() function we have
	const int trcsz = zrg_.width()+1;
	trc_ = new SeisTrc( trcsz );

So the size of 'trc_' depends on zrg_. But this can be different cube_.getZSz()
which can be larger than zrg_. And in that case we were writing values beyond 
the array leading to crash.

Test Plan: No more crash when Z range subselection is done.

Reviewers: yuancheng, helene

Reviewed By: helene

Subscribers: kristofer, helene, raman

Differential Revision: https://phabricator.dgbes.com:8443/D6885

git-svn-id: https://svn.opendtect.org/od/trunk@37937 a985855d-d384-4b45-ae33-31a410998c3c