od: Fix for resizing volume box in wheeler scene
Summary:
Came across this while working on Mantis 3430.
The Z range of Z axis transformed scene can be completely different from that of the survey.
Hence calling SI().snapZ() on this zrange doesn't work as expected. 
And the volume's zrange width becomes zero in this scenario.
Hence moved the call in else block when 'datatransform_' is null.

Test Plan: Resizing volume box in wheeler scene works.

Reviewers: jaap, kristofer

Reviewed By: kristofer

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D6987

git-svn-id: https://svn.opendtect.org/od/trunk@38059 a985855d-d384-4b45-ae33-31a410998c3c