od5.0: Added Pos::IdxPair2Coord::isSubsetOf()
Summary:
Check, not perfect but not bad either, whether an IdxPair2Coord object supports
nodes that are part of another IdxPair2Coord.
This test may not be water-tight but in pratice I think it will be good enough. At least, if it fails then the two are not compatible.

Test Plan: Used in Petrel connection seems to work fine.

Reviewers: helene

Reviewed By: helene

Differential Revision: https://phabricator.dgbes.com:8443/D7036

git-svn-id: https://svn.opendtect.org/od/branches/od5.0@38079 a985855d-d384-4b45-ae33-31a410998c3c