od: bug fixing 3739 about well text size ratio in screen snapshot
Summary:
This revision is same as D6860.
The problem is caused by setting dpidensity to these texts. since we set the text character size mode as text::object in well display, in order to change
the text's size by zooming. But once we reset the dpi to the text, the size of text will be changed in the text::object mode.after doing
screen snapshot, the size is not the size on the screen any more.

Test Plan:
Compiles ok, and Tested by make different dpi screen snapshot, the result is expected. No matter dpi is changed,
the text size always is same as the size on the screen.

Reviewers: kristofer

Reviewed By: kristofer

Subscribers: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D6897

git-svn-id: https://svn.opendtect.org/od/trunk@38087 a985855d-d384-4b45-ae33-31a410998c3c