od: Fix for saving Fault/FaultStickSet color
Summary:
http://bugs.dgbes.com/view.php?id=3847
Fault/FaultStickSet color is not saved. The changes made in D5397 and D5731 were
only applicable in context of Horizons. In case of horizons we save two files 
.hor and .par files. The .hor files have all the data and info related to horizon.
The additional info related to horizons like regional markers were stored in 
.par file ( see Mantis 2937 for details ).

So the previous changes were done assuming that every EMObject has the data and
.par file. So an easy fix would be to write this file for every EMObject. 
In that case we can remove this constraint from EMManager::writeDisplayPars() function ioobj is horizon. 

bool EMManager::writeDisplayPars(..) 
{
    IOObjInfo ioobjinfo( mid );
    if( !ioobjinfo.isOK() || !ioobjinfo.isHorizon() )
	return false;

    .....
    .....
	
} 


So we could change this to 

bool EMManager::writeDisplayPars(..) 
{
    IOObjInfo ioobjinfo( mid );
    if( !ioobjinfo.isOK() )
	return false;

    .....
    .....
}

Or else we could continue the way things were and only write par files
in case of horizons. In that case we may go with changes done in this revision. 

What are your suggestions?

Test Plan: Saving fault/faultstickset color works.

Reviewers: ding, nanne

Reviewed By: nanne

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D7038

git-svn-id: https://svn.opendtect.org/od/trunk@38109 a985855d-d384-4b45-ae33-31a410998c3c