od: Avoiding returning pointer to an invalid string
Summary:
factoryDisplayName() returns a uiString. The returned object is deleted
immediately when function returns from name(). 
getOriginalString() returns the pointer of data in the uiString, but as the
uiString object has been deleted the pointer is no longer valid.
getFullString returns the pointer of a static string. Hence it is safe.
An alternateive to this would be to return factoryKeyword() which returns a const char*.

Test Plan: The user interface of Batch programs look OK after the change.

Reviewers: kristofer, bert

Reviewed By: bert

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D7123

git-svn-id: https://svn.opendtect.org/od/trunk@38188 a985855d-d384-4b45-ae33-31a410998c3c