od5.0-horizon display improvement
Summary:
This revision is same as revision as D6995 in od related to Mantis issue 3793 and 3776. The only differences are
in handling with ABI consistency in interfaces.

Test Plan:
Compiles fine. Tests with od5.0 in with same horizon, the speed is improved a lot.
And also tested with all horizon features including creating, tracking, and so on. it works as
expectation.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D7002

git-svn-id: https://svn.opendtect.org/od/branches/od5.0@38311 a985855d-d384-4b45-ae33-31a410998c3c