od: New feature on the Basemap plugin: 'Add Contours'
Summary:
Now contours can be added to the basemap.

It's needed to load the horizon to have de z range preadjusted on the spacing field. Otherwise, the field will be blank.

Test Plan:
It was tested on the F3_Data and compared against the contours already stored in F3_Data.
The contours were generated in the same way that stored contours in the F3_Data.

Reviewers: nanne

Reviewed By: nanne

Maniphest Tasks: T149

Differential Revision: https://phabricator.dgbes.com:8443/D7190

git-svn-id: https://svn.opendtect.org/od/trunk@38365 a985855d-d384-4b45-ae33-31a410998c3c