od: Fix for memory leak
Summary:
As Seis::ParallelReader's constructor is overloaded, its optional to pass the arrays as an argument.
When the second constructor is called ParallelReader(const IOObj&,const TrcKeyZSampling&); 
it creates its own array by allocating memory and this memory is never freed. Hence made changes 
to keep a track of allocated memory and delete it in the destructor.

Test Plan:
Seis::ParallelReader is used in new Sub Volume Finder. Tested by running SVF and stepping through 
the code and keeping track of memory use in TaskManager.

Reviewers: nanne, kristofer

Reviewed By: kristofer

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D7219

git-svn-id: https://svn.opendtect.org/od/trunk@38368 a985855d-d384-4b45-ae33-31a410998c3c