od- memory leak in vishorizondisplay
Summary:
This is the same revision as D7060 in od5.0
The memory leak is detected by Intel inspector XE about the usage of BufferStringSet.
I have found the "replace" calling which has also missed "delete".

Test Plan:
Compiles fine. Tests by Inspector, the leak is gone.

About vitileresolutiondata, I have fixed the memory leaking issue, after test again, the leaking are gone too. But the changes should be in the revision D7002.
So for the time being, I leave it locally.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D7061

git-svn-id: https://svn.opendtect.org/od/trunk@38373 a985855d-d384-4b45-ae33-31a410998c3c