od-horizon display improvement
Summary:
This revision is related to Mantis issue 3793 and 3776.
when current od version loads and displays a huge big horizon, its speed is obviously slower than od4.6. I have done some investigation about the reasons as below:
1. Reading speed.
2. Visualization speed.

The visualization speed is mainly caused by
1.) set positions to tile resolution data.
2.) normal computation.
3.) too much vertices in render( over 100 million vertices )
4.) texture channel ( attribute mapping to color )
5.) using visCoordinates class storing huge number of positions.

In this revision,
1.) changed some part of designs in both HorizonTile and TileResolution data. The main ideas are using same vertices, texture coordinates, and normals
for all kinds of resolution data to the same tile. So the resolution data class only need to store primitive set tessellated by vertices and resolution.  
This can save a lot of memory and also render time.
2.) using osg::vec3arry instead of visCoordinates to store tile's vertices. This can reduce the consuming time in setting position.
3.) about normal computation. we only need to compute once for one tile if horizon geometry is not changed. so move the individual vertex normal computation
    to the same loop in setposition function in the class HorizonTile. thus we can save some time. once geometry is changed we do only update single vertex normal.
4.) removed all invalid normal issues, because we have initial normals if the computed nomal is invalid, we are not going to set it to vertex.

The all changes made in this revision are around above ideas.

Test Plan:
Compiles fine. Tests with od5.0 in with same horizon, the speed is improved a lot.
And also tested with all horizon features including creating, tracking, and so on. it works as 
expectation.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D6995

git-svn-id: https://svn.opendtect.org/od/trunk@38374 a985855d-d384-4b45-ae33-31a410998c3c