Bugfix: far inl/crl/z slice was disappearing when zooming in.
Summary:
Bugfix: far inl/crl/z slice was disappearing when zooming in. Only happened
when (1) plane was perpendicular to viewing direction, (2) positioned on one
of the farthest lines of the survey, and (3) the zooming makes the front of
another object (e.g. the survey box) disappear behind the camera eye, while
its back is still in front of the camera eye. Same happens in OD4.6, so it
is not OSG related. It neither is a TexturePlane issue, because it happens
with other flat objects as well, draw e.g. a flat fault plane over there.
The OSG nearFarRatio is 0.0005 by default, while the value to solve this is
almost constant (somewhere between 0.001-0.0015) for all the surveys I tried.
It might have something to do with depth buffer resolution, but I can't explain
why it does not just cause z-fighting and instead removes a whole object from
the scene. The consequence of increasing nearFarRatio to 0.002 is that objects
disappear a bit earlier when zooming it on them, but this still happens far
beyond the resolution of their finest details. Unfortunately, I found no proof
yet that it will solve other Mantis issues on objects disappearing from screen.

Test Plan:
Tested for multiple surveys, big/small, oblong/square. Add a section on one
of the farthest lines of the survey, press the view inl/crl/z button, and
start zooming in, either with or without survey box or a resized perpendicular
section occurring in the front part of the scene only. Can now zoom in without 
disappearing section. Next add another section on one of the closest lines of
the survey to verify that one can still extremely zoom in on it before it
disappears behind the camera eye.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D7240

git-svn-id: https://svn.opendtect.org/od/trunk@38455 a985855d-d384-4b45-ae33-31a410998c3c