od5.0- Added addPolygonOffset in vismarkerset class
Summary:
This revision is the same as D7320 in od.
Sometimes the markers are hidden by other objects in display, such as in well display, they are hidden by logs( Paul found this problem).
The solution here we added a polygonoffset in vismarkerset. it works fine

Test Plan: Compiles and Tests fine

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D7321

git-svn-id: https://svn.opendtect.org/od/branches/od5.0@38536 a985855d-d384-4b45-ae33-31a410998c3c