od5.0: Fix for Select button on top of combobox
Summary:
In the FingerPrint attribute the Select button is on top of the combobox, instead of to the right.

Real problem is that these uiObjects inside a uiTable are never finalized, i.e. the finalise() function is not called.
Real solution is therefor needed in uiTable, but this fix works for now.

Test Plan: Show attribute editor and select Fingerprint attribute. Button in uiAttrSel is again where it should be.

Reviewers: helene

Reviewed By: helene

Differential Revision: https://phabricator.dgbes.com:8443/D7332

git-svn-id: https://svn.opendtect.org/od/branches/od5.0@38549 a985855d-d384-4b45-ae33-31a410998c3c