od5.0: Make copy seismics possible in batch
Summary:
BG needs a tool to copy the i.* cubes they made to the final cubes.
od_copy_seis may be that thing.
Spin-off is that copying cubes can now be done in batch.

Test Plan: Tested in od and od5.0 for a couple of cubes/line data sets.

Reviewers: raman

Reviewed By: raman

Maniphest Tasks: T250

Differential Revision: https://phabricator.dgbes.com:8443/D7386

git-svn-id: https://svn.opendtect.org/od/branches/od5.0@38605 a985855d-d384-4b45-ae33-31a410998c3c