od: Added log file support to class JobCommunic to enable log output for process communication status
Summary: Any process which will use this class to communicate with the od_MMBatch, for multimachine processing environment will have the ability to create a log file on the local machines disk to put all the communication states in that file.

Test Plan:
Tested on a single machine environment, and forced the process to fail.
All the socket communication staus was loged in the file, including the success and failure status.

Reviewers: nanne

Reviewed By: nanne

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D7400

git-svn-id: https://svn.opendtect.org/od/trunk@38651 a985855d-d384-4b45-ae33-31a410998c3c