od5.0: Removed SEGYDirect option from PS3D output fields
Summary: http://bugs.dgbes.com/view.php?id=4090 SEGYDirect write is not available for Prestack 3D data. Therefore they are disabled now. The only exception is the Inserters where you do select a SEGYDirect object as 'output' but do not do the actual writing. This is easily handled by making allownonuserselectable_=true for the output field in the Inserter.

Test Plan: I am still able to insert SEGYDirect objects for PS3D but it does not appear as an option while writing, for example in 'Copy' tool.

Reviewers: bert

Reviewed By: bert

Differential Revision: https://phabricator.dgbes.com:8443/D7417

git-svn-id: https://svn.opendtect.org/od/branches/od5.0@38657 a985855d-d384-4b45-ae33-31a410998c3c