Bugfix: "Set Annotation Scale" dialog did not function properly (Mantis 3720)
Summary:
Bugfix: "Set Annotation Scale" dialog did not function properly (Mantis 3720).
Looks like two developers responsible for dialog and scene did not understand
each other.
(1) Dialog was initialized with empty TraceKeyZSampling instead of the default
axis scale of the survey box that is displayed in that case. 
(2) The stop values returning from the dialog were neither processed nor used.

Test Plan:
Tested by filling in lots of different values in the dialog fields and
monitoring whether the dialog values and the axis labels in the scene
correspond with each other.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D7425

git-svn-id: https://svn.opendtect.org/od/trunk@38702 a985855d-d384-4b45-ae33-31a410998c3c