od5.0: Fix for wrong wavelet in initial welltie window
Summary: the wavelet was not forwarded from RaySynthGenerator to MultiTraceSynthGenerator to SynthGenerator if the wavelet object was different but with the same name as a wavelet from the database.

Test Plan: tested in F3_Demo. The synthetic trace now has the right frequency content without having the push "additional information button"

Reviewers: satyaki, bert

Reviewed By: bert

Subscribers: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D7428

git-svn-id: https://svn.opendtect.org/od/branches/od5.0@38727 a985855d-d384-4b45-ae33-31a410998c3c