od5.0: Added a sleep of 1 sec after launching a background process
Summary: I do not have a theoretical base for this, it is more of an empirical fix as adding these statements fixes the MMP failure we had with DTECT_DEBUG flag on. Basically the launching shell stays alive for another second just to give enough time for the child process to initialize.

Test Plan: Tested only in 5.0 release environment. Without this change MMP fails with DTECT_DEBUG on. After this fix, MMP works irrespective of the value of DTECT_DEBUG

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D7455

git-svn-id: https://svn.opendtect.org/od/branches/od5.0@38732 a985855d-d384-4b45-ae33-31a410998c3c