od: Set uiAttrSel as readonly
Summary:
Read bug http://bugs.dgbes.com/view.php?id=2445
Making the field readonly solves the problem for now.

Better might be to get the input text, and if it is stored data, add it to the descset, similar to what the uiAttrSelDlg does.

Test Plan: Opened attribute set editor. Pasting text in uiAttrSel fields is not possible anymore.

Reviewers: helene

Reviewed By: helene

Subscribers: helene

Differential Revision: https://phabricator.dgbes.com:8443/D7464

git-svn-id: https://svn.opendtect.org/od/trunk@38746 a985855d-d384-4b45-ae33-31a410998c3c