od: Changed the atsampleposition's check statement
Summary:
Changed the condition that checks whether a value is at a sample position or not as well as the limit value to this check (1e-6 to 1e-4).

I was having problems when trying to display some z slices on my Basemap so @Nanne found out that true atsamplepos values were being set as false (that is to say, getting into the conditional statement and being set as false).

Test Plan:
Now z-slices can be diplayed on Basemaps without crashing.

Before, everytime that I was trying to display a z-slice, od was crashing because some invalid memory access reading.

Reviewers: helene

Reviewed By: helene

Subscribers: helene, nanne

Differential Revision: https://phabricator.dgbes.com:8443/D7506