od: Implemented a tool for edititng path/name of SEGY files in a SEGYDirect datastore.
Summary:
This sounds simple but I would like to mention a few points:
1. I am using the 'browse' button in 'Manage Seismics' window and I could not find a way to change the icon of the button if the current object is SEGYDirect.
2. Instead of writing the whole sgydef file I am only editing the Footer that contains the file data. I hope this is safe. If the new Footer is smaller than the old, I pad up with spaces.

Test Plan: Created a SEGYDirect datastore with 2 files. Moved the files to another location. Renamed the files. For every change, I used this tool to edit the sgydef file and displayed the data on a Z slice. All looks OK.

Reviewers: bert

Reviewed By: bert

Maniphest Tasks: T249

Differential Revision: https://phabricator.dgbes.com:8443/D7538