od6.0-bug fixing: remove points from horizon causes "scene gone"
Summary:
This revision is same as revsion D7547 in od.
This is bug report from Nanne. when remove horizon tracking point and doing zoom in/out somtimes the scene is gone.
After investigation, the problem is caused by hideFromDisplay function()

Changes make it works.

Test Plan: Compiles and tests fine

Reviewers: kristofer

Reviewed By: kristofer

Subscribers: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D7555