od6.0 bug fixing: intersection line between fault and horizon
Summary:
This is same revision as D7583 in od.

This bug is already existing since we have this function in od. The problem was caused by uncertain order of intersections which are
computed by triangle plane and horizon. The function addAndSortToResult() can not complete making a correct point order in output line if
horizon is not flat. That is why the intersection line between fault and z-slice is correct, because all points are on a straight line (but the line has repeat drawing).
However the function addAndSortToResult can find out the start point. it is useful to this solution.

Solution: added two extra functions to make a correct order which is used to connect points as one line.

Test Plan: Complies fine. Tests with F3-demo and one other survey working fine

Reviewers: yuancheng

Reviewed By: yuancheng

Subscribers: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D7587