od bug 3539 fixing
Summary:
I found the current version does not work well in below:
1. when shifting  horizon and then change the z-scale, the shifting translation is not applied in z-scale. because the z-scale is included in new display translationform.
2. since we introduce new osgnode "vistransform" once we have shifting operation on horizon, once we have this node, the rest of operation like add/remove object not working any more.

This revision fixed the bug 3539 and also potential bug which will be caused by the introduction of visTransform in shifting.

Test Plan:
Compiles fine. Tests is fine with multi horizons in operation of adding, shifting and removing, and adding/removing inline, crossline and z-slice to see the changes of intersection lines.
All works.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D7606