od: Added inl/crl option to Basemap [gui only]
Summary:
Inl/Crl were added to the tree.

Just the gui were made yet in order to let the tree looking like the od tree.

Test Plan: It pops up a message while you try to add either a in-line or cross-line item to Basemap saying that it's not fully implemented yet.

Reviewers: nanne

Reviewed By: nanne

Maniphest Tasks: T149

Differential Revision: https://phabricator.dgbes.com:8443/D7716