od5.0: Fix for memory leak in Horizon Gridding
Summary:
http://bugs.dgbes.com/view.php?id=4287 The gridder was never deleted. This was affecting the DipSteered gridding the most as the Steering data was not released.

Making the destructor virtual should be OK for ABI compatibility. The pre-compiled plugins will still consider the destructor as non-virtual and will still have the memory leak. But OD and the dGB plugins will not have the leak anymore.

Test Plan: Tested by putting breakpoints in the destructors of the specific gridder classes. They are called now and in case of Dip-stered gridding there is a visual change in memory consumption after gridding is finished.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D7797