Bugfix: Horizon contour labels occluded when viewed from top (Mantis 3970).
Summary:
(1) Bugfix: Horizon contour labels were occluded when viewed from top. This
is inherent to displaying text in rotate-to-screen mode. Solved by an angular
elevation with respect to a reference plane, which has been implemented in
osgGeo. This is a real improvement, because the problem existed in OpendTect
with Coin3D too, and was not just introduced by OSG. 
(2) Extended the horizon contour properties dialog with options to change
font (size), left-right alignment and elevation angle. This gives the user
maximum flexibility to make all labels visible if the defaults are not good
enough for specific viewing angles and/or rough local horizon curvature.

Test Plan:
Tested on a not-too-flat horizon with contour labels. The default elevation
angle keeps labels visible at all useful camera positions without any serious
distortion. Changing the elevation angle, font size or left-right alignment
of the labels in the properties dialog produces an immediate change on screen.
New settings are remembered but not saved, being mainly used to create the
perfect picture in a specific situation. Tested for perspective/orthographic
projection, and for surveys with reversed inl-crl sense as well.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D7805