od 6.0- Griding with algorithm of Continuous Curvature
Summary:
This is the same revision as D7536 in od trunk

The added algorithm is based on GMT's surface.cc and other related functions in GMT package.
The idea of this algorithm is using a tension to control surface's trend in which it controls from
whole surface to blocks, and from blocks to grids. it is an iteration process based on the theoretic
method mentioned on their paper.

This revision implements also an method of initializing grid by an given radius. for the time being
we don't have it in od UI. we can add it if we want. the default radius is 0.

This revision has not implement fault (break line) yet. it is on going. But I think firstly we need
improve current od's method of computing intersection line between fault and horizon.

I found the parameter of convergence limitation is very important. it directly influence the result.
the value is smaller, the computing time longer. but the result is better. The default value is defined as
0.001x(Z difference in the area of horizon ). We don't have this parameter in od UI. do we need it?

Other changes are related the consistency between factory user name and key name. They influence the what are displayed in UI.

Added test application

Test Plan: Compiles fine and Tested fine by test application and F3 demo in od.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D7704