od: Enable "Remove slection" button only when selection mode is active.
Summary:
The "Remove Selection" button remains always active, even when there is no selction by the polygon selector, and there is actually nothing to remove, and the icon is also same as that of the "Remove from tree" button which makes it more confusing, to find the actuall purpose of this button.

Changed as a request from Lakshmi, this time the buton will be active only if the polygon selction tool is active.

Test Plan: Works as expected

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D7941