od6.0: Included withCheckBox option to uiSelNrRange class
Summary: Included option to have a check box beside a uiSelNrRange that allow the user enable or disable the its fields. It is just option, therefore it shouldn't do any changes on implementations that already use this class.

Test Plan:
Tested checking the uiSelNrRange field in the HorizonCube dialog (Survey>Pre-load>Load 3D)- nothing changed as expected.
Plus, tested on uiBasemapSeismicOutline dialog with the withCheckBox option enabled and everything looked as planned.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D7946