Bugfix: CmdRecorder missed listbox double click that closes dialog. (PORTING FROM od TO od6.0 UNCHANGED)
Summary:
Bugfix: CmdRecorder missed listbox double click that closes dialog. Reported
by Hardeep, and already existing since CmdRecorder was developed. In case
the code contains a callback that accepts/rejects the dialog in which the
triggering object resides, the object will disappear before its possibly
pending command(s) are flushed. Now flushing is forced if the windowClosed
notifier of its source window is triggered. Note that the close notifier of
uiObject is not useful in this case. This is triggered in the destructor
of the base class, and the required dynamic casting to the derived class
(e.g. uiListBox) is not possible anymore at that time.

Test Plan:
Tested in F3_Demo by recording user action 'TreeMenu "In-line" "Add"' followed
by 'ListClick "Select Data" "5 ChimneyCube" DoubleLeft'. Now recorded script
contains double-click command, and running it works fine too.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D7838