od5.0: Fixed skipping of first Inline while reading 3D horizons
Summary: http://bugs.dgbes.com/view.php?id=4382. The HiddenParam arrparam is initialized in createSection() which is called while reading the first inline. So, re-extracting the value of arr from arrparam after the createSection call solves it.

Test Plan: Tested reading of a version 1 and a version 3 horizon. All inlines are read now. Also in the context of the Mantis bug, gridding includes all inlines from the input horizon.

Reviewers: satyaki, nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D7970