od5.0-improvement on intersection line between fault and horizon
Summary:
This revision is same as D8074.
This revision comes from mantis issue 4162. Although the previous revision improved the intersection line algorithm, it still
has problems when the terrain of horizon is complicated. In this revision, it has done below aspects:

1, Removed the existing function: addAndSortToResult(), because sometimes it sorts the un-correct start point,
for example, if d00<d01<d11<d10, it will do line 323(original code).
2. For current design, we only want to find out a start point of the intersection line. It is suppose that a fault can not be a circle. we use function findMin() to find the start point having min X or min Y. if all X is almost same ( within a limitation)
than using min Y. At the same time, the function checks if the start point is located in the middle of line in one direction. if it is in the middle of X, then it will find min Y
point as start point.
3. Removed the neighbour points which are very closed to a point. usually they are duplicated points or they have only difference in Z.
In a certain xy range, we want to keep only one point, thus we can reduce the error in making points order. On the other hand, we have enough points to make
3D intersection line.
4. optimize point order of the intersection. After sort points to a line. it still could have error such as a point located in a wrong order in the intersection line.
we using XY extent to check and remove the error from the point list. in line 419, we don't use bbox.includes() function, because there is an equal condition in that function.

Test Plan: Compiles fine. Tests by loading more than 15 horizons and Fault A, B, C in F3-Demo-d55 survey, they work fine.

Reviewers: yuancheng

Reviewed By: yuancheng

Subscribers: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D8081