od5.0: Fixed the auto-loading of some plugins like matlab
Summary: loading of the MATLABLink plugin fails because the function SharedLibAccess::errMsg() was declared but not implemented. Also added filters and default folder for the shared-libraries file selection dialog, and added auto-load of the selected file.

Test Plan: plugin shows up. Interface is correct. However the volume builder step does not run.

Reviewers: nanne, helene

Reviewed By: helene

Differential Revision: https://phabricator.dgbes.com:8443/D8097